safari-534.53-branch/LayoutTests/fast/css/counters/2displays-expected.txt

This test passes if both the first and the second counter are 1000.

PASS layoutTestController.counterValueForElementById('testView') is '1000 1000'

TEST COMPLETE
